Create module to shorten encryption library class
