+++ braille.txt
@@ -1,0 +1,6 @@
+.000..000.0...000.00..0.0.0.00...00..0.0...00.0.00..000.0...0.0.0...000.000....0
+0.0....0.0.........0..00.0...0..00000.0...0000.0.0...0.0......00.0...0.0.0.0..00
+......000.00......0...0.........0.....0...0.....0...000.00....0.......0.0......0
+.0.00....0.00.0.0..0.00.00...00.0.0.0.
+0.0000..0.00.000..000..0.0..000000.0.0
+..0.......0...0...0...0.0...0...0.....
